🧹 refactor: flatten out file structure of db module
- request sub-module was too big and did too many things.
- broke it down into (params, response, request) submodules.
- moved getting db time and db connection to a DbConn struct.
- moved DbTable and associated structs/implementations to
  table submodule.
- added commends and documnatation to several functions/methods.
