+++ src/db/params/tests/user_params.rs
@@ -4,7 +4,7 @@ async fn init_test(conn: &SqlitePool) -> Result<Vec<StoreUser>> {
     Ok(try_join_all(TEST_USERS.into_iter().map(async |user_name| {
         Ok::<StoreUser, Error>({
             UserParams::builder()
-                .user_name(ParamOption::new().map_value(user_name).clone())
+                .with_user_name(ParamOption::new().map_value(user_name).clone())
                 .build()
                 .post(conn)
                 .await?
@@ -50,7 +50,7 @@ async fn test_get_user(conn: SqlitePool) -> Result<()> {
     try_join_all(init_test(&conn).await?.into_iter().map(async |want| {
         anyhow::Ok::<()>({
             let got = UserParams::builder()
-                .user_id(ParamOption::new().map_value(want.id).clone())
+                .with_user_id(ParamOption::new().map_value(want.id).clone())
                 .build()
                 .get(&conn)
                 .await?;
@@ -69,7 +69,7 @@ async fn test_get_user(conn: SqlitePool) -> Result<()> {
 async fn test_delete_user(conn: SqlitePool) -> Result<()> {
     let original = init_test(&conn).await?;
     let params = UserParams::builder()
-        .user_id(
+        .with_user_id(
             ParamOption::new()
                 .map_value(original.get(0).unwrap().id)
                 .clone(),
@@ -105,8 +105,8 @@ async fn test_update_user(conn: SqlitePool) -> Result<()> {
         .zip(want)
         .map(|(user, name)| {
             UserParams::builder()
-                .user_id(ParamOption::new().map_value(user.id).clone())
-                .user_name(ParamOption::new().map_value(name).clone())
+                .with_user_id(ParamOption::new().map_value(user.id).clone())
+                .with_user_name(ParamOption::new().map_value(name).clone())
                 .build()
         })
         .collect::<Vec<UserParams>>();
@@ -143,7 +143,7 @@ async fn test_update_user(conn: SqlitePool) -> Result<()> {
 async fn test_invalid_params(conn: SqlitePool) -> Result<()> {
     let no_id = UserParams::builder().build();
     let no_name = UserParams::builder()
-        .user_id(ParamOption::new().map_value(0).clone())
+        .with_user_id(ParamOption::new().map_value(0).clone())
         .build();
 
     match no_id.delete(&conn).await {
