+++ src/db/totals.rs
@@ -1,14 +1,20 @@
 use std::sync::Mutex;
-use tokio::sync::OnceCell;
 
 use super::*;
 
+pub mod prelude {
+    pub use super::StoreTotal;
+}
+
 static TOTALS: OnceCell<Mutex<StoreTotal>> = OnceCell::const_new();
+
+#[derive(Debug, Default, PartialEq)]
+pub struct StoreTotal(HashMap<i64, Decimal>);
 
 impl StoreTotal {
     // setters
     async fn new() -> Result<Self> {
-        let conn = get_db().await?;
+        let conn = DbConn::try_get().await?;
 
         let rows_to_calc =
             sqlx::query_file_as!(StoreJoinRaw, "sql/get_ru_no_offset.sql")
@@ -73,5 +79,47 @@ impl StoreTotal {
             .get(&key)
             .copied()
             .ok_or(anyhow::Error::msg("No total found for given key."))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::db::params::tests::init_join_rows;
+
+    fn intermediate_totals() -> Vec<HashMap<i64, Decimal>> {
+        vec![
+            HashMap::from([(3, dec!(9.98))]),
+            HashMap::from([(2, dec!(9.49))]),
+            HashMap::from([(2, dec!(8.32)), (3, dec!(8.32))]),
+        ]
+    }
+    fn expected_totals() -> HashMap<i64, Decimal> {
+        HashMap::from([(3, dec!(18.30)), (2, dec!(17.81))])
+    }
+
+    #[sqlx::test]
+    async fn test_get_totals(conn: SqlitePool) -> Result<()> {
+        init_join_rows(&conn).await?;
+        let want = expected_totals();
+        let mut got = StoreTotal::default();
+
+        JoinedReceiptParams::builder()
+            .with_offset(0)
+            .build()
+            .get_all(&conn)
+            .await?
+            .into_iter()
+            .zip(intermediate_totals())
+            .try_for_each(|(row, want)| {
+                anyhow::Ok({
+                    assert_eq!(want, row.try_calc()?);
+                    got.add(row.try_calc()?);
+                })
+            })?;
+
+        assert_eq!(want, got.0, "Test if all the math is right ✨");
+
+        Ok(())
     }
 }
