+++ src/table/render.rs
@@ -13,7 +13,7 @@ impl TableData {
     pub(super) fn render_rows(&self, style: Style) -> Vec<Row> {
         self.items
             .iter()
-            .map(|data| data.ref_array().style(style))
+            .map(|data| data.into_row().style(style))
             .collect()
     }
